Move ChannelContextType to core module
