+++ saleor/graphql/channel/types.py
@@ -1,12 +1,10 @@
 import collections
 import itertools
-from typing import TYPE_CHECKING, TypeVar, cast
+from typing import TYPE_CHECKING
 
 import graphene
-from django.db.models import Model
 from django_countries.fields import Country
 from graphene.types.objecttype import ObjectType
-from graphene.types.resolver import get_default_resolver
 from promise import Promise
 
 from ...channel import models
@@ -19,7 +17,6 @@ from ...permission.enums import (
 )
 from ..account.enums import CountryCodeEnum
 from ..core import ResolveInfo
-from ..core.context import ChannelContext
 from ..core.descriptions import (
     ADDED_IN_318,
     ADDED_IN_320,
@@ -39,7 +36,6 @@ from ..core.scalars import Day, Hour, Minute
 from ..core.types import BaseObjectType, CountryDisplay, ModelObjectType, NonNullList
 from ..meta.types import ObjectWithMetadata
 from ..tax.dataloaders import TaxConfigurationByChannelId
-from ..translations.resolvers import resolve_translation
 from ..warehouse.dataloaders import WarehousesByChannelIdLoader
 from ..warehouse.types import Warehouse
 from .dataloaders import ChannelWithHasOrdersByIdLoader
@@ -51,56 +47,6 @@ from .enums import (
 
 if TYPE_CHECKING:
     from ...shipping.models import ShippingZone
-
-T = TypeVar("T", bound=Model)
-
-
-class ChannelContextTypeForObjectType(ModelObjectType[T]):
-    """A Graphene type that supports resolvers' root as ChannelContext objects."""
-
-    class Meta:
-        abstract = True
-
-    @staticmethod
-    def resolver_with_context(
-        attname, default_value, root: ChannelContext, info: ResolveInfo, **args
-    ):
-        resolver = get_default_resolver()
-        return resolver(attname, default_value, root.node, info, **args)
-
-    @staticmethod
-    def resolve_id(root: ChannelContext[T], _info: ResolveInfo):
-        return root.node.pk
-
-    @staticmethod
-    def resolve_translation(
-        root: ChannelContext[T], info: ResolveInfo, *, language_code
-    ):
-        # Resolver for TranslationField; needs to be manually specified.
-        return resolve_translation(root.node, info, language_code=language_code)
-
-
-class ChannelContextType(ChannelContextTypeForObjectType[T]):
-    """A Graphene type that supports resolvers' root as ChannelContext objects."""
-
-    class Meta:
-        abstract = True
-
-    @classmethod
-    def is_type_of(cls, root: ChannelContext[T] | T, _info: ResolveInfo) -> bool:
-        # Unwrap node from ChannelContext if it didn't happen already
-        if isinstance(root, ChannelContext):
-            root = root.node
-
-        if isinstance(root, cls):
-            return True
-
-        if cls._meta.model._meta.proxy:
-            model = root._meta.model
-        else:
-            model = cast(type[Model], root._meta.model._meta.concrete_model)
-
-        return model == cls._meta.model
 
 
 class StockSettings(BaseObjectType):
