+++ saleor/graphql/shipping/mutations/shipping_price_remove_product_from_exclude.py
@@ -4,8 +4,8 @@ import graphene
 
 from ....permission.enums import ShippingPermissions
 from ....shipping import models
-from ...channel.types import ChannelContext
 from ...core import ResolveInfo
+from ...core.context import ChannelContext
 from ...core.doc_category import DOC_CATEGORY_SHIPPING
 from ...core.mutations import BaseMutation
 from ...core.types import NonNullList, ShippingError
